3차 미완성본

+++ src/main/java/com/sparta/library/repository/RentRepository.java
@@ -7,5 +7,5 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface RentRepository extends JpaRepository<Rent, Long> {
+public interface RentRepository extends JpaRepository<Rent, Integer> {
 }
